feat: playlist track test 코드 작성
