+++ src/main/java/com/team/comma/domain/PlaylistTrack.java
@@ -4,12 +4,12 @@ import jakarta.persistence.*;
 import lombok.*;
 
 @Entity
-@Data
+@Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "playlist_track_tb")
-public class Playlist_Track extends BaseEntity{
+public class PlaylistTrack extends BaseEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
