+++ src/main/java/com/team/comma/domain/Track.java
@@ -3,6 +3,8 @@ package com.team.comma.domain;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.util.List;
+
 @Entity
 @Getter
 @Builder
