+++ src/main/java/com/team/comma/repository/PlaylistTrackRepository.java
@@ -1,0 +1,13 @@
+package com.team.comma.repository;
+
+import com.team.comma.domain.PlaylistTrack;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface PlaylistTrackRepository extends JpaRepository<PlaylistTrack, Long> {
+
+    List<PlaylistTrack> findAllByPlaylist_Id(Long id);
+    List<PlaylistTrack> findAllByTrack_Id(Long id);
+
+}
