+++ src/main/java/com/team/comma/repository/TrackRepository.java
@@ -1,10 +1,11 @@
 package com.team.comma.repository;
 
+import com.team.comma.domain.PlaylistTrack;
 import com.team.comma.domain.Track;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface TrackRepository extends JpaRepository<Track, Long> {
-    List<Track> findAllByPlaylistId(Long id);
+
 }
