+++ src/test/java/com/team/comma/service/PlaylistTrackServiceTest.java
@@ -1,0 +1,4 @@
+package com.team.comma.service;
+
+public class PlaylistTrackServiceTest {
+}
